+++ packages/eval-suite/src/index.ts
@@ -1,4 +1,5 @@
 export * from "./types.js";
+export * from "./contentPublishingDogfoodEvalCases.js";
 export * from "./dogfoodEvalCases.js";
 export * from "./enterpriseDogfoodEvalCases.js";
 export * from "./evalCases.js";